requirements.txt
azure-cosmos==4.5.1
python-dotenv==1.0.0
